Split CardMovement class into its own file
