+++ Assets/Scripts/CardMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b64aea4d5dbfdc04f860f1dc681f3154
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
